+++ RELEASES.md
@@ -45,6 +45,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template: updated mongo storage to Bindings, replacing deprecated BindingsBean
 * prov-storage-mongodb: updated mongo storage to Bindings, replacing deprecated BindingsBean
 * service-translator, service-templates: introduced storage configuration file
+* prov-storage-mongodb, prov-jsonld: fixed issue with handling of bundle Id and date options
 
 
 `2.0.5`